Normalize top-level APIs
