+++ docs/source/reference/api/eland.DataFrame.es_info.rst
@@ -1,0 +1,10 @@
+eland.DataFrame.es_info
+=======================
+
+.. currentmodule:: eland
+
+.. warning::
+    Previously this method was named ``info_es()``.
+    DataFrame.info_es() is deprecated, use DataFrame.es_info() instead.
+
+.. automethod:: DataFrame.es_info
