--- docs/source/reference/api/eland.DataFrame.info_es.rst
@@ -1,6 +1,0 @@
-eland.DataFrame.info_es
-=======================
-
-.. currentmodule:: eland
-
-.. automethod:: DataFrame.info_es
