+++ docs/source/reference/api/eland.DataFrame.to_pandas.rst
@@ -1,0 +1,6 @@
+eland.DataFrame.to_pandas
+=========================
+
+.. currentmodule:: eland
+
+.. automethod:: DataFrame.to_pandas
