+++ docs/source/reference/api/eland.Series.es_info.rst
@@ -1,0 +1,10 @@
+eland.Series.es_info
+====================
+
+.. currentmodule:: eland
+
+.. warning::
+    Previously this method was named ``info_es()``.
+    Series.info_es() is deprecated, use Series.es_info() instead.
+
+.. automethod:: Series.es_info
