--- docs/source/reference/api/eland.Series.info_es.rst
@@ -1,6 +1,0 @@
-eland.Series.info_es
-====================
-
-.. currentmodule:: eland
-
-.. automethod:: Series.info_es
