+++ docs/source/reference/api/eland.Series.to_pandas.rst
@@ -1,0 +1,6 @@
+eland.DataFrame.to_pandas
+=========================
+
+.. currentmodule:: eland
+
+.. automethod:: Series.to_pandas
