+++ docs/source/reference/api/eland.csv_to_eland.rst
@@ -1,0 +1,6 @@
+eland.csv_to_eland
+==================
+
+.. currentmodule:: eland
+
+.. autofunction:: csv_to_eland
