--- docs/source/reference/api/eland.read_csv.rst
@@ -1,6 +1,0 @@
-eland.read_csv
-==============
-
-.. currentmodule:: eland
-
-.. autofunction:: read_csv
