--- docs/source/reference/api/eland.read_es.rst
@@ -1,6 +1,0 @@
-eland.read_es
-=============
-
-.. currentmodule:: eland
-
-.. autofunction:: read_es
