+++ eland/tests/common.py
@@ -24,10 +24,10 @@ from eland.tests import (
 _pd_flights = pd.read_json(FLIGHTS_DF_FILE_NAME).sort_index()
 _pd_flights["timestamp"] = pd.to_datetime(_pd_flights["timestamp"])
 _pd_flights.index = _pd_flights.index.map(str)  # make index 'object' not int
-_ed_flights = ed.read_es(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME)
+_ed_flights = ed.DataFrame(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME)
 
 _pd_flights_small = _pd_flights.head(48)
-_ed_flights_small = ed.read_es(ES_TEST_CLIENT, FLIGHTS_SMALL_INDEX_NAME)
+_ed_flights_small = ed.DataFrame(ES_TEST_CLIENT, FLIGHTS_SMALL_INDEX_NAME)
 
 _pd_ecommerce = pd.read_json(ECOMMERCE_DF_FILE_NAME).sort_index()
 _pd_ecommerce["order_date"] = pd.to_datetime(_pd_ecommerce["order_date"])
@@ -37,7 +37,7 @@ _pd_ecommerce["products.created_on"] = _pd_ecommerce["products.created_on"].appl
 _pd_ecommerce.insert(2, "customer_birth_date", None)
 _pd_ecommerce.index = _pd_ecommerce.index.map(str)  # make index 'object' not int
 _pd_ecommerce["customer_birth_date"].astype("datetime64")
-_ed_ecommerce = ed.read_es(ES_TEST_CLIENT, ECOMMERCE_INDEX_NAME)
+_ed_ecommerce = ed.DataFrame(ES_TEST_CLIENT, ECOMMERCE_INDEX_NAME)
 
 
 class TestData:
@@ -68,7 +68,7 @@ def assert_pandas_eland_frame_equal(left, right):
         raise AssertionError(f"Expected type ed.DataFrame, found {type(right)} instead")
 
     # Use pandas tests to check similarity
-    assert_frame_equal(left, right._to_pandas())
+    assert_frame_equal(left, right.to_pandas())
 
 
 def assert_eland_frame_equal(left, right):
@@ -79,7 +79,7 @@ def assert_eland_frame_equal(left, right):
         raise AssertionError(f"Expected type ed.DataFrame, found {type(right)} instead")
 
     # Use pandas tests to check similarity
-    assert_frame_equal(left._to_pandas(), right._to_pandas())
+    assert_frame_equal(left.to_pandas(), right.to_pandas())
 
 
 def assert_pandas_eland_series_equal(left, right, check_less_precise=False):
@@ -90,4 +90,4 @@ def assert_pandas_eland_series_equal(left, right, check_less_precise=False):
         raise AssertionError(f"Expected type ed.Series, found {type(right)} instead")
 
     # Use pandas tests to check similarity
-    assert_series_equal(left, right._to_pandas(), check_less_precise=check_less_precise)
+    assert_series_equal(left, right.to_pandas(), check_less_precise=check_less_precise)
