+++ eland/tests/dataframe/test_init_pytest.py
@@ -20,15 +20,15 @@ class TestDataFrameInit:
 
         # Construct invalid DataFrame (throws)
         with pytest.raises(ValueError):
-            ed.DataFrame(client=ES_TEST_CLIENT)
+            ed.DataFrame(es_client=ES_TEST_CLIENT)
 
         # Construct invalid DataFrame (throws)
         with pytest.raises(ValueError):
-            ed.DataFrame(index_pattern=FLIGHTS_INDEX_NAME)
+            ed.DataFrame(es_index_pattern=FLIGHTS_INDEX_NAME)
 
         # Good constructors
         ed.DataFrame(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME)
-        ed.DataFrame(client=ES_TEST_CLIENT, index_pattern=FLIGHTS_INDEX_NAME)
+        ed.DataFrame(es_client=ES_TEST_CLIENT, es_index_pattern=FLIGHTS_INDEX_NAME)
 
         qc = QueryCompiler(client=ES_TEST_CLIENT, index_pattern=FLIGHTS_INDEX_NAME)
-        ed.DataFrame(query_compiler=qc)
+        ed.DataFrame(_query_compiler=qc)
