+++ eland/tests/field_mappings/test_scripted_fields_pytest.py
@@ -43,7 +43,7 @@ class TestScriptedFields(TestData):
 
         # note 'None' is printed as 'NaN' in index, but .index shows it is 'None'
         buf = StringIO()
-        ed_field_mappings.info_es(buf)
+        ed_field_mappings.es_info(buf)
         print(buf.getvalue())
 
         expected = self.pd_flights().columns.to_list()
