+++ eland/tests/series/test_rename_pytest.py
@@ -23,14 +23,14 @@ class TestSeriesRename(TestData):
         print(pd_renamed)
         print(ed_renamed)
 
-        print(ed_renamed.info_es())
+        print(ed_renamed.es_info())
 
         assert_pandas_eland_series_equal(pd_renamed, ed_renamed)
 
         pd_renamed2 = pd_renamed.rename("renamed2")
         ed_renamed2 = ed_renamed.rename("renamed2")
 
-        print(ed_renamed2.info_es())
+        print(ed_renamed2.es_info())
 
         assert "renamed2" == ed_renamed2.name
